requirements.txt
streamlit==1.39.0
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-pinecone==0.2.0
pinecone==5.3.1
openai==1.40.0
python-dotenv==1.0.0
pypdf==5.1.0
httpx==0.27.2